fix: fixing build error
Error: You cannot use different slug names for the same dynamic path ('id' !== 'slug').
